Create geojson file, need to transform lambert to lon/lat with specific CC by departement
